+++ services/scenario_service.py
@@ -1,3 +1,5 @@
+# scenario_service.py
+
 import streamlit as st
 
 from agents.scenario_generator import (
