+++ utilities/setup_utils.py
@@ -1,6 +1,20 @@
-import streamlit as st
+# setup_utils.py
 
-from utilities.utils import set_api_keys, setup_logging
+import logging
+import os
+
+import streamlit as st
+from dotenv import load_dotenv
+from langfuse.callback import CallbackHandler
+
+
+def setup_logging():
+    """Sets up logging configuration."""
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s - %(levelname)s - %(message)s",
+        handlers=[logging.StreamHandler()],
+    )
 
 
 def setup_page():
@@ -31,3 +45,22 @@ def load_css():
     with open(".css/app_styles.css", "r") as f:
         css = f.read()
     st.markdown(f"<style>{css}</style>", unsafe_allow_html=True)
+
+
+def set_api_keys(env_file_path=".env"):
+    """Loads and sets necessary API keys for OpenAI and LangFuse."""
+    logging.info("Attempting to load API keys from specified .env file.")
+    load_dotenv(env_file_path)
+
+    openai_key = os.getenv("OPENAI_API_KEY")
+    if openai_key:
+        logging.info("OpenAI API key loaded successfully.")
+
+    langfuse_keys = ["LANGFUSE_PUBLIC_KEY", "LANGFUSE_SECRET_KEY", "LANGFUSE_HOST"]
+    if all(os.getenv(key) for key in langfuse_keys):
+        logging.info("LangFuse API keys loaded successfully.")
+
+
+def setup_langfuse_keys(pk: str, sk: str, host: str) -> CallbackHandler:
+    """Creates a LangFuse CallbackHandler with the provided keys."""
+    return CallbackHandler(public_key=pk, secret_key=sk, host=host)
