--- utilities/utils.py
@@ -1,53 +1,0 @@
-# utils.py``
-
-import json
-import logging
-import os
-from pathlib import Path
-
-from dotenv import load_dotenv
-from langfuse.callback import CallbackHandler
-
-
-def set_api_keys(env_file_path=".env"):
-    """Loads and sets necessary API keys for OpenAI and LangFuse."""
-    logging.info("Attempting to load API keys from specified .env file.")
-    load_dotenv(env_file_path)
-
-    openai_key = os.getenv("OPENAI_API_KEY")
-    if openai_key:
-        logging.info("OpenAI API key loaded successfully.")
-
-    langfuse_keys = ["LANGFUSE_PUBLIC_KEY", "LANGFUSE_SECRET_KEY", "LANGFUSE_HOST"]
-    if all(os.getenv(key) for key in langfuse_keys):
-        logging.info("LangFuse API keys loaded successfully.")
-
-
-def setup_langfuse_keys(pk: str, sk: str, host: str) -> CallbackHandler:
-    """Creates a LangFuse CallbackHandler with the provided keys."""
-    return CallbackHandler(public_key=pk, secret_key=sk, host=host)
-
-
-def setup_logging():
-    """Sets up logging configuration."""
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(asctime)s - %(levelname)s - %(message)s",
-        handlers=[logging.StreamHandler()],
-    )
-
-
-def save_uploaded_file(uploaded_file, save_dir="/tmp"):
-    """Saves an uploaded file to the specified directory and returns the path."""
-    try:
-        save_directory = Path(save_dir)
-        save_directory.mkdir(parents=True, exist_ok=True)
-        file_path = save_directory / uploaded_file.name
-
-        with open(file_path, "wb") as f:
-            f.write(uploaded_file.getvalue())
-
-        return file_path
-    except Exception as e:
-        logging.error(f"Failed to save uploaded file: {str(e)}")
-        raise
